Windows CI using AppVeyor
enable 64 bit Julia too

don't fetch --unshallow if repo is complete

This check is necessary if there are fewer commits in the history
than the depth of the AppVeyor clone (currently 50).

DataFrames is well over 50 commits, I'm putting this in for the
sake of other packages that might look at this file as an example.

update Julia nightly download location

Update AppVeyor configuration to use stable

and disable email notifications (hopefully)