Fix InexactError in join() due to too many levels
PooledDataVecs() uses the smallest possible integer size, but
the DataFrames code stores higher references than that in the vector.
Always use UInt32, which will be enough for any reasonable number of
levels.

Maybe PooledDataVecs() should not compact its result by default
(which makes it type-unstable), but this should be considered as part
of a larger rework.