Loosen constructor for a DataFrame
It seems to be a little more difficult than it should be to pass a container of columns to the `DataFrame`'s constructor. This change lets you, e.g.

```julia
DataFrame([[1,2,3]], [:a])
```

Before this would complain because `[[1,2,3]]` is of type `Vector{Vector{Int}}` which is not a subtype of (concrete) type `Vector{Any}` (type parameters aren't convariant). This will automatically loosen the type parameter for the column container, if necessary.