Avoid calling vcat(dfs...) in combine()
This avoids compiling specialized methods for each number of arguments,
which can be very large when using groupby(). The code does not take
advantage at all of that information anyway. The code prior to
refactoring used to do that, by exporting
vcat(::Vector{AbstractDataFrame}). Use an internal method instead since
this is non-standard.

Incidentally, this fixes a small bug introduced in the refactoring: when
working with an empty data frame, vcat() was called, and it returns
Any[0], which surprisingly implied the addition of a bogus x1 column to
the result. The test was actually correct before refactoring.