Override getproperty, setproperty! and propertynames for AbstractDataFrame and DataFrameRow
Provides a nice shorthand syntax for simple getindex and setindex! uses, with autocompletion,
and makes DataFrameRow more like NamedTuple.
This requires replacing all direct field accesses with calls to accessor functions.
Do not mention this feature in the manual yet since it is not available on a released Julia
version. Most uses of df[:col] will better be changed to df.col when Julia 0.7 is released.