Fix a bug in unstack(df)
Currently `unstack(df)` defaults to `:id` being a rowkey. This is not documented, and actually against documentation (which says that we take all columns except `colkey` and `value` as `rowkeys`) and not intuitive (why `:id` should be special?). Therefore I fix it as a bug.

However, this is breaking as in some cases a resulting data frame will contain more columns than before (it is not very painful I guess because we produce more not less but still breaking).

Therefore we might also go through deprecation period (or change the behavior but print a warning if we produce more than `:id` column in this case). Not sure what to do