Use Tables.allocatecolumn to simplify and improve code
This new interface allows removing special cases for CategoricalArray and
making the code more general. It also makes it easier to stop calling
promote_col_type(data...) which triggers a stack overflow when
the number of input arguments is high.

This relies on promotion rules which have been fixed in CategoricalArrays 0.3.14,
and which give a slightly different behavior: now combining a CategoricalArray with
an Array gives an Array, since the number of unique levels could be very high.