Deprecate delete!, insert! and merge!
Rename delete! to deletecols! and insert! to insertcol! to fit the definition of data frames
as collections of rows. This will allow renaming deleterows! to delete! after the deprecation
period.

Remove merge! since it is trivial to write by hand, and it is likely to confuse users coming
from languages where joins are performed using function "merge".