Add fast grouping method for CategoricalArray keys
When grouping on a single CategoricalArray column, we don't need a Dict to get group codes.
This approach cannot be used when joining, as the Dict is needed.
group_rows now handles the sorting, which allows skipping this step if the grouping
is guaranteed to be sorted (as is the case for CategoricalArray).
This mechanism can be extended in the future to handle multiple CategoricalArray keys,
or integer keys (using radix sort).