Deprecate delete!, insert! and merge! (#1560)
Rename delete! to deletecols! and insert! to insertcols! to fit the definition of data frames
as collections of rows. This will allow renaming deleterows! to delete! after the deprecation
period.

Remove merge! since it is trivial to write by hand, and it is likely to confuse users coming
from languages where joins are performed using function "merge".

* Deprecate df[col] = nothing in favor of deletecols!(df, col)

This is an R-ism which doesn't make sense in Julia and isn't more convenient
than the more standard function call form.

* Fix insertcol! bug when nrow(df) == 0

The code probably intended to check whether there were no columns.

* Improve insertcols! API using keyword argument and pair

This makes it more similar DataFrame constructors and to push! for dicts,
and the order of arguments is more natural and simpler to remember.