Use faster hashing approach for first CategoricalVector grouping key (#1565)
Since hash() takes as an argument the hash for the previous keys, we cannot in general
reuse precomputed hashes. However, this is possible for the first grouping key since the
previous hash is always zero in that case. Since the number of keys is generally small,
it makes a large difference to improve performance for this case.