Generalize optimized categorical grouping method to multiple keys (#1674)
The compiler is still able to generate efficient code specialized on the
number of keys. Also require groups !== nothing for this method since
we never call it without groups and it cannot compute the actual
number of nonempty groups without doing all the work anyway.

Also fix a bug with the groups field with sort=true or skipmissing=true.
Make tests stricter to catch it.