Deprecate colwise
This function does not make a lot of sense since it doesn't preserve column nor function names.
The GroupedDataFrame method loses grouping information andreturns a vector of vectors, which
isn't very useful. It also conflicts with Distances.colwise.

aggregate is a much more useful replacement, and the current behavior of colwise can easily be
obtained using a comprehension, which is even shorter now that we support eachcol.