Speed up getindex with AbstractVector{Bool} row selection
Computing the integer indices before indexing into the vectors is faster.
The array indexing code uses a LogicalIndex wrapper which computes the number
of true indices and doesn't allocate a vector of integer indices, but it's
only slightly faster when there's a single column, and slower for more than
one column.