Stop using makeunique=true for grouping keys in combine
Instead, check whether columns with the same names as grouping keys
have been returned, and throw an error if they are not equal.
This is breaking since columns named key_1, etc. won't be created,
but in practice it should not be frequently used. Introducing a deprecation
would be more annoying than helpful since it would require adding a keyword
argument to disable adding grouping keys and force users to set it to silence
the warning, which happens in most common cases.