Stop reexporting packages
Doing this automatically includes all functions exported by CategoricalArrays, Missings
and InvertedIndices to our API. Instead, export selected functions that we really use
and/or overload. This has the most significant consequences for CategoricalArrays
(`levels`, `isordered`, `CategoricalArray`...). The advantage is that this will allow
us to drop this dependency if we want (e.g. by defining `categorical`, `allowmissing!`, etc.
in a lightweight interface package).