Enhance docs with concept introduction, how to reference single cells, using CSV
Added a section at the beginning explaining the concept of a DataFrame. 
Added a section on how to reference single cells (prior to referencing groups of cells)
Revised the use of CSV.jl in the IO section. 

One suggestion was to replace the DataFrame picture I used with an actual screenshot of Julia. I deferred that until later, due to:
* Wanted something colorful at the top 
* Future doc revisions will replace it (see below) -- so it's colorful placeholder 

Future - the documerntation needs to be reworked from top to bottom, using a single, compact DataFrame that be used for all of the example code. The documentation currently creates a lot of ad-hoc, small DataFrames with column names like "A", "B", "x1", "x2", etc., which are confusing and creates visual clutter. 

I'll need to invent that single, compact DataFrame that can be used to illustrate all of the concepts. Perhaps two will be needed.