Change `stack` to create a `PooledArray{String}` column by default
This is less likely to confuse users (it's the dplyr and pandas default),
and it would allow dropping the dependency on CategoricalArrays if we wanted to.
The drawback of returning strings is that the order of old columns is only
reflected by the order of rows, and no longer by the sorting order of levels.

Also return a `PooledArray{Symbol}` when `variable_eltype=Symbol` is passed.
That's more consistent and more efficient, e.g. if you want to group on symbols
later.

Any element type is now accepted, with `CategoricalValue{String}` being handled
by the general path. In practice it would allow using a custom `AbstractString`
type, though that's relatively unlikely.