Add using(julia) and import(python)
Some pandas examples uses `pd.` and `np.` prefixes
so that it's shorter. This commit adds the import
statement accordingly.

I also added the corresponding `using` statement in
the Julia sample.  The `mean` function comes from
`Statistics` so I added that also to avoid unnecessary
surprises when the readers try to followin the examples.