[BREAKING] Deprecate `categorical` and `categorical!` (#2394)
These methods are the only remaining traces of CategoricalArrays in the public API.
Getting rid of them will allow dropping the dependency if we want.

`transform` and `transform!` now provide quite reasonable replacements for some uses.
What is lacking though is a way to select columns based on their types, which would
be a useful addition in general. Also it could make sense to add a keyword argument
to keep original names, which would be convenient to avoid repeating the selector.