RFC/DNM: profile-guided despecialization
This reduces specialization of numerous method arguments.
The list was generated by automated tools and may not be
particularly well-chosen, as it depends on the nature of the workload.
This was generated by running the test suite, which is probably
not representative of real-world workloads.

The list is sorted from low-to-high inference time, so the ones towards
the end are more important.

This shaves about 200s off the time to run the tests (out of 700s total).
Similar changes to Base & other dependencies would shave off more,
though always with the risk of runtime penalty.