added show(df) for a better view of DFs
I think this is one that we have kind of needed for some time -- stylable, simple HTML output. As of right now, I am realizing that I am going to have to loop eachrow of the df since the html is easier to populate by row. Let me fix that!